Fixed administration form rendering bug with English language.
The translation files for the English language were not consistant with the
referenced keys in the blueprints, leading to array-to-string conversions. The
files were updated to use the same structure and the blueprint was modified
to match the English files, as it was better.

Probably an unfinished change that was commited erroneously.

Fixes #1.